Ajustement des liens d'assets et placement des images sur le Login et le SideMenu
