+++ src/components/login/Login.css
@@ -14,7 +14,7 @@
     -o-border-radius: 25px 25px 175px 175px;
 }
 
-.LogoIcoTop{
+.LogoIcoLogin{
     width: 300px;
     height: 300px;
     border-radius: 175px;
