+++ src/components/login/Login.jsx
@@ -42,7 +42,7 @@ function LoginPage() {
   return (
     <div className="LoginPage grid grid-rows-1 place-content-center">
       <div className="LoginPageBox border-4 border-white m-auto p-4 text-center bg-gray-200 rounded-xl shadow-xl">
-        <div className="LogoIcoTop items-center m-5 shadow-md">
+        <div className="LogoIcoLogin items-center m-5 shadow-md">
           <img src={logo} alt="logo_ImmoCo" className="LogoIcoPicture" />
         </div>
         <div className=" grid grid-rows-4">
