--- src/components/navigation/Navbar.css
@@ -1,19 +1,0 @@
-nav{
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    min-height: 10vh ;
-    background-color: aquamarine;
-    color: white;
-  }
-.nav-links{
-    width: 40%;
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    list-style: none;
-  }
-.nav-space{
-  margin-right: 10px;
-  color: black;
-}
